refactor(Capital name): Edit names directory
