+++ src/App.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import Router from './Router';
 import Snackbar from './components/Snackbar';
-import { SnackbarContextProvider } from './contexts/SnackbarContext';
-import { ProyectContextProvider } from './contexts/ProyectContext';
-import { UserContextProvider } from './contexts/UserContext';
+import { SnackbarContextProvider } from './hooks/SnackbarContext';
+import { ProyectContextProvider } from './hooks/ProyectContext';
+import { UserContextProvider } from './hooks/UserContext';
 import { createMuiTheme } from '@material-ui/core/styles';
 import { ThemeProvider } from '@material-ui/styles';
 const theme = createMuiTheme({
