+++ src/components/ColorSetter.js
@@ -2,7 +2,7 @@ import 'rc-color-picker/assets/index.css';
 import React, { useState, useContext } from 'react';
 import Grid from '@material-ui/core/Grid';
 import { Panel as ColorPickerPanel } from 'rc-color-picker';
-import { ColorContext } from '../contexts/ColorContext';
+import { ColorContext } from '../hooks/ColorContext';
 import ColorPicker from 'rc-color-picker';
 import Box from '@material-ui/core/Box';
 import Button from '@material-ui/core/Button';
