+++ src/components/PaletteList.js
@@ -10,10 +10,10 @@ import Collapse from '@material-ui/core/Collapse';
 import ExpandLess from '@material-ui/icons/ExpandLess';
 import ExpandMore from '@material-ui/icons/ExpandMore';
 import paletteListModel from '../utils/paletteListModel';
-import { ColorContext } from '../contexts/ColorContext';
-import { initAxiosInterceptors } from '../Helpers/auth-helper';
+import { ColorContext } from '../hooks/ColorContext';
+import { initAxiosInterceptors } from '../helpers/auth-helper';
 import Axios from 'axios';
-import { ProyectContext } from '../contexts/ProyectContext';
+import { ProyectContext } from '../hooks/ProyectContext';
 import Select from '@material-ui/core/Select';
 import MenuItem from '@material-ui/core/MenuItem';
 import TextField from '@material-ui/core/TextField';
