+++ src/components/ShareCard.js
@@ -3,10 +3,10 @@ import Grid from '@material-ui/core/Grid';
 import Paper from '@material-ui/core/Paper';
 import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
-import { ProyectContext } from '../contexts/ProyectContext';
-import { SnackbarContext } from '../contexts/SnackbarContext';
+import { ProyectContext } from '../hooks/ProyectContext';
+import { SnackbarContext } from '../hooks/SnackbarContext';
 import Axios from 'axios';
-import { initAxiosInterceptors } from '../Helpers/auth-helper';
+import { initAxiosInterceptors } from '../helpers/auth-helper';
 
 initAxiosInterceptors();
