+++ src/views/Dashboard.js
@@ -12,12 +12,12 @@ import {
   getCurrentUser,
   deleteToken,
   initAxiosInterceptors
-} from '../Helpers/auth-helper';
+} from '../helpers/auth-helper';
 
-import { ColorProvider } from '../contexts/ColorContext';
-import { ChatContextProvider } from '../contexts/ChatContext';
-import { ProyectContext } from '../contexts/ProyectContext';
-import { UserContext } from '../contexts/UserContext';
+import { ColorProvider } from '../hooks/ColorContext';
+import { ChatContextProvider } from '../hooks/ChatContext';
+import { ProyectContext } from '../hooks/ProyectContext';
+import { UserContext } from '../hooks/UserContext';
 
 initAxiosInterceptors();
